% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{data.dump}
\alias{data.dump}
\title{Data Dump
A function to check the kind of data that needs to be packed and sent.
Different configurations of the data will be enclosed with different named lists.}
\usage{
data.dump(fields, size, data = NULL)
}
\arguments{
\item{fields}{}

\item{size}{}

\item{data}{}
}
\value{

}
\description{
Data Dump
A function to check the kind of data that needs to be packed and sent.
Different configurations of the data will be enclosed with different named lists.
}
