% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.BAclient.download}
\alias{.BAclient.download}
\title{Basic Authentication HTTP download with get function
This function will download files from the Qd server and store it in the path provided
This is a (hidden) internal function}
\usage{
.BAclient.download(file_name, path, ...)
}
\arguments{
\item{file_name}{the file that you would like to download from the Qd server}

\item{path}{the location wehre the file needs to be downloaded}

\item{...}{}
}
\value{
http operation code
}
\description{
Basic Authentication HTTP download with get function
This function will download files from the Qd server and store it in the path provided
This is a (hidden) internal function
}
