% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.BAclient.get}
\alias{.BAclient.get}
\title{Basic Authentication HTTP get function
The get function will get the requested page in JSON format from the
Qd server
This is a (hidden) internal function}
\usage{
.BAclient.get(page_name, ...)
}
\arguments{
\item{page_name}{\itemize{
\item name of the page to get
}}

\item{...}{}
}
\value{
http operation code
}
\description{
Basic Authentication HTTP get function
The get function will get the requested page in JSON format from the
Qd server
This is a (hidden) internal function
}
\examples{
.BAclient.get("/example")
}
