% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.BAclient.patch}
\alias{.BAclient.patch}
\title{Basic Authentication HTTP patch function
The patch function will take the page name and the data that needs to be patched.
In return you will get http code
This is a (hidden) internal function}
\usage{
.BAclient.patch(page_name, data, ...)
}
\arguments{
\item{page_name}{\itemize{
\item name of the page
}}

\item{data}{\itemize{
\item JSON data to be updated on the page.
}}

\item{...}{}
}
\value{
http operation code
}
\description{
Basic Authentication HTTP patch function
The patch function will take the page name and the data that needs to be patched.
In return you will get http code
This is a (hidden) internal function
}
