% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.BAclient.unload}
\alias{.BAclient.unload}
\title{Basic Authentication HTTP delete function
This function will delete the page on the Qd server. It requires the page name
This is a (hidden) internal function}
\usage{
.BAclient.unload(page_name, ...)
}
\arguments{
\item{page_name}{name of the page to be removed from the Qd server}

\item{...}{}
}
\value{
http operation code
}
\description{
Basic Authentication HTTP delete function
This function will delete the page on the Qd server. It requires the page name
This is a (hidden) internal function
}
\examples{
.BAclient.unload("/example")
}
