% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.BAclient.upload}
\alias{.BAclient.upload}
\title{Basic Authentication HTTP upload with post function
This function will upload the files provided in the fle list to the Qd server
This is a (hidden) internal function}
\usage{
.BAclient.upload(files_list, ...)
}
\arguments{
\item{files_list}{list of files to be uploaded}

\item{...}{}
}
\value{
http operation code, and list of file locations on the Qd server
}
\description{
Basic Authentication HTTP upload with post function
This function will upload the files provided in the fle list to the Qd server
This is a (hidden) internal function
}
\examples{
file_list <- list("aml.jpg","dai.jpg")
.BAclient.upload(file_list)
}
