% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.delta_name_change}
\alias{.delta_name_change}
\title{Make a required name change if needed}
\usage{
.delta_name_change(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Make a required name change if needed
}
