% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.get.env.var}
\alias{.get.env.var}
\title{Search and return an existing, or default environmental variable
This is a (hidden) internal function}
\usage{
.get.env.var(object)
}
\arguments{
\item{object}{An environmental variable}
}
\value{
The found or default environmental variable \code{object}
}
\description{
Search and return an existing, or default environmental variable
This is a (hidden) internal function
}
\examples{
test_variable <- 'INTERNAL_ADDRESS'
.get_env_var(test_variable)
}
