% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.is_primitive}
\alias{.is_primitive}
\title{Function to check if the object is a primitive.
This function checks if the object is of type numeric, integer, logical, character.
This is a (hidden) internal function}
\usage{
.is_primitive(object = NULL)
}
\arguments{
\item{object}{}
}
\value{
TRUE - logical
}
\description{
Function to check if the object is a primitive.
This function checks if the object is of type numeric, integer, logical, character.
This is a (hidden) internal function
}
\examples{
test_variable <- 1
.is_primitive(test_variable)
}
