% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Configure the environmental variables
This is a (hidden) internal function}
\usage{
.onLoad(...)
}
\value{

}
\description{
The environmental variables configured here are:
internal_address
external_address
hub_address
hub_access_key_id
hub_access_key_secret
shutdown_timeout
They are stored in a list \code{.config}
}
