% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.page.diff}
\alias{.page.diff}
\title{Function to provide the difference between two instances of a page.
The existing and the new instances of the page.
The returning data is passed to the page.save() function
This is a (hidden) internal function}
\usage{
.page.diff(ref_page, new_page)
}
\arguments{
\item{ref_page}{existing page}

\item{new_page}{updated page
Adding a list of lists required (temporarily broken)}
}
\value{
.page_changes returns the differences in page
}
\description{
Function to provide the difference between two instances of a page.
The existing and the new instances of the page.
The returning data is passed to the page.save() function
This is a (hidden) internal function
}
\examples{
.page.diff(.test_page, test_page)
}
