% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{.site.save}
\alias{.site.save}
\title{Function to send a page on a site to the Qd server
The function requires the name of the page and the data to be sent
This is a (hidden) internal function}
\usage{
.site.save(page_name, data, ...)
}
\arguments{
\item{page_name}{name of the page}

\item{data}{\itemize{
\item JSON data that needs to be sent
}}

\item{...}{}
}
\value{
return code from sending the page to the server
}
\description{
Function to send a page on a site to the Qd server
The function requires the name of the page and the data to be sent
This is a (hidden) internal function
}
\examples{
.site.save("/example",json_data)
}
