% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durham_method.R
\name{durham_ve}
\alias{durham_ve}
\title{Estimate VE using method from Durham et al. 1988 (AJE)}
\usage{
durham_ve(x, df = 2, n_days, n_periods, n_days_period, var)
}
\arguments{
\item{x}{survival object}

\item{df}{degrees of freedom}

\item{n_days}{number of days in the study}

\item{n_periods}{number of periods in the study}

\item{n_days_period}{number of days per period}

\item{var}{name of variable to assess}
}
\value{
tibble of VE estimates for each period (the estimate for each period is the average VE over the days
within each period.
}
\description{
This method is based on smoothing scaled residuals from the Cox proportional hazard regression model.
It consists of four steps. First, an ordinary proportional hazard model is fitted using a partial likelihood
function. Second, Schoenfeld residuals are calculated. These residuals are used to test the independence
between residuals and time. Third, the residuals are scaled and added to the coefficient from the ordinary
proportional hazard regression model. Fourth, after smoothing we can get the estimated hazard ratios as
function of time by recovering the time-varying regression coefficient. This allows testing the hypothesis of
no VE waning and the estimation of TVE at each time point.
}
\keyword{wave}
