% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_methods.R
\name{ferdinands_ve}
\alias{ferdinands_ve}
\title{Estimate VE using method from Ferdinands et al. 2017}
\usage{
ferdinands_ve(dat, splines = FALSE)
}
\arguments{
\item{dat}{data set.}

\item{splines}{logical. whether or not to include calendar time as a spline.}
}
\value{
estimates of the likelihood ratio under each model and the difference between the two models.
}
\description{
Estimate VE using method from Ferdinands et al. 2017
}
\keyword{wave}
