% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{glist}
\alias{glist}
\title{Glist - Creating a guarded list.
This list will be guarded and wont be changed.}
\usage{
glist(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Glist - Creating a guarded list.
This list will be guarded and wont be changed.
}
