% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{likelihood_xxx}
\alias{likelihood_xxx}
\title{Likelihood function for maximum likelihood (ML) method}
\usage{
likelihood_xxx(x, alpha, theta_0, phi)
}
\arguments{
\item{x}{a list with the information necessary to calculate the likelihood function}

\item{alpha}{alpha parameter}

\item{theta_0}{theta_0 parameter}

\item{phi}{phi parameter}
}
\value{
the negative sum of the loglikelihood function
}
\description{
The ML method is based on calculating the contribution to the likelihood function of each study participant.
See XX for more details.
}
\keyword{wave}
