% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_method.R
\name{logLik}
\alias{logLik}
\title{Likelihood function for maximum likelihood (ML) method}
\usage{
logLik(x, par)
}
\arguments{
\item{x}{a list with the information necessary to calculate the likelihood function}

\item{par}{parameters to estimate}
}
\value{
the negative sum of the loglikelihood function
}
\description{
The ML method is based on calculating the contribution to the likelihood function of each study participant.
See XX for more details.
}
\keyword{wave}
