% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_method.R
\name{ml_ve}
\alias{ml_ve}
\title{Estimate VE using a maximum likelihood (ML) method}
\usage{
ml_ve(
  dat,
  n_days,
  n_periods,
  n_days_period,
  latent_period = 1,
  infectious_period = 4
)
}
\arguments{
\item{dat}{data set}

\item{n_days}{number of days in the study}

\item{n_periods}{number of periods in the study}

\item{n_days_period}{number of days per period}

\item{latent_period}{length of latent period}

\item{infectious_period}{length of infectious period}
}
\value{
list with a tibble of VE estimates for each period (the estimate for each period is the average VE over the days
within each period) and the maximu likelihood estimates of the parameters.
}
\description{
This method is similar to that of Ainslie et al. 2017 (SIM). The ML method is based on calculating the
contribution to the likelihood function of each study participant. See XX for more details.
}
\keyword{wave}
