% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_method_w_mcmc.R
\name{ml_ve2}
\alias{ml_ve2}
\title{Estimate VE using a maximum likelihood (ML) method with MCMC}
\usage{
ml_ve2(data, params, my_prior = NULL, file_name = "test", par_tab, mcmc_pars)
}
\arguments{
\item{data}{data set}

\item{params}{input parameters}

\item{my_prior}{prior function for MCMC}

\item{file_name}{character string of output file name}

\item{par_tab}{data frame of parameter values}

\item{mcmc_pars}{vector of MCMC options}
}
\value{
list with a tibble of VE estimates for each period (the estimate for each period is the average VE over the days
within each period) and the maximu likelihood estimates of the parameters.
}
\description{
This method is similar to that of Ainslie et al. 2017 (SIM). The ML method is based on calculating the
contribution to the likelihood function of each study participant. See XX for more details.
}
\keyword{wave}
