% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{my_creation_function}
\alias{my_creation_function}
\title{Likelihood function for maximum likelihood (ML) method}
\usage{
my_creation_function(parTab, data, PRIOR_FUNC, ...)
}
\arguments{
\item{parTab}{The parameter table controlling information such as bounds, initial values etc.}

\item{data}{The data frame of data to be fitted.}

\item{PRIOR_FUNC}{Prior distribution function. If no priors are specified (PRIOR_FUNC = NULL), then the uniform prior of the sampling restrictions will apply from parTab.}

\item{...}{other function arguments}
}
\value{
the negative sum of the loglikelihood function
}
\description{
The ML method is based on calculating the contribution to the likelihood function of each study participant.
See XX for more details.
}
\keyword{wave}
