% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{page.add}
\alias{page.add}
\title{Function adds one or many cards to the page_instance.
Each of these cards contains information that will be displayed.
These cards could be static of dynamic.
Each card requires a unique name. If a card with a name similar to an existing card
is added then the existing card will be over written.
The card name cannot be empty.}
\usage{
page.add(page_instance, page_name, card_name, FUN, ...)
}
\arguments{
\item{page_instance}{\itemize{
\item The page instance object
}}

\item{page_name}{\itemize{
\item name of the page instance
}}

\item{card_name}{name of the card}

\item{FUN}{card function to be passed. This could be any card type available in ui.R}

\item{...}{}
}
\value{
page_instance - list()
}
\description{
Function adds one or many cards to the page_instance.
Each of these cards contains information that will be displayed.
These cards could be static of dynamic.
Each card requires a unique name. If a card with a name similar to an existing card
is added then the existing card will be over written.
The card name cannot be empty.
}
