% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{page.load}
\alias{page.load}
\title{Function to load/get the data on a page in JSON format
A page name needs to be passed to the function to get back the
data on a the page in JSON page.}
\usage{
page.load(page_name, ...)
}
\arguments{
\item{page_name}{\itemize{
\item name of the page
}}

\item{...}{}
}
\value{
page in the form of a JSON
}
\description{
Function to load/get the data on a page in JSON format
A page name needs to be passed to the function to get back the
data on a the page in JSON page.
}
\examples{
page.load("/example")
}
