% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{page}
\alias{page}
\title{Function to create a new page or check the existence of a page
This function is used to create a new page on a site.
If the name of the page is an empty string then the function returns an error.
If the name of the page does not have a \code{/} it returns a error.
If another function calls the page function (by setting fun_flag = TRUE) then a new page is not created
but the existence of a page is checked. Under this condition of the page does not exist, then the function
returns an error.
\code{fun_flag} is a variable that is used when another function calls this \code{page}
function. It could}
\usage{
page(page_name, fun_flag = FALSE, ...)
}
\arguments{
\item{page_name}{\itemize{
\item the name of the page to be created, or to be checked
}}

\item{fun_flag}{\itemize{
\item a flag set when other functions call this function.
}}

\item{...}{}
}
\value{
page_instance - list()
}
\description{
Function to create a new page or check the existence of a page
This function is used to create a new page on a site.
If the name of the page is an empty string then the function returns an error.
If the name of the page does not have a \code{/} it returns a error.
If another function calls the page function (by setting fun_flag = TRUE) then a new page is not created
but the existence of a page is checked. Under this condition of the page does not exist, then the function
returns an error.
\code{fun_flag} is a variable that is used when another function calls this \code{page}
function. It could
}
