% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{page.save}
\alias{page.save}
\title{Function to save the page, push the page to the Qd server
The function requires the page instance created, and the name of the page.}
\usage{
page.save(page_instance, page_name, ...)
}
\arguments{
\item{page_instance}{\itemize{
\item page instance object
}}

\item{page_name}{\itemize{
\item name of the page
}}

\item{...}{}
}
\value{
return code from sending the page to the server
}
\description{
Function to save the page, push the page to the Qd server
The function requires the page instance created, and the name of the page.
}
