% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{page_frame}
\alias{page_frame}
\title{Page Frame
Add additional attributes if the page is having a delta frame.}
\usage{
page_frame(page, ...)
}
\arguments{
\item{page}{}

\item{...}{}
}
\value{

}
\description{
Page Frame
Add additional attributes if the page is having a delta frame.
}
