% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParams.R
\name{readParams}
\alias{readParams}
\title{Read in parameters}
\usage{
readParams(filename = NULL)
}
\arguments{
\item{filename}{path to csv file}
}
\value{
list of parameters
}
\description{
Function reads in parameters for simulation from a csv file
}
\keyword{wave}
