% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{site.download}
\alias{site.download}
\title{This function will download file(s) from that Qd server to local disk.
The file will be written to the location defined in the path variable}
\usage{
site.download(files_name, path, ...)
}
\arguments{
\item{files_name}{name of the file to be downloaded from the Qd server}

\item{path}{folder location where the file needs to be downloaded}

\item{...}{}
}
\value{
file is written to disk and http code
}
\description{
This function will download file(s) from that Qd server to local disk.
The file will be written to the location defined in the path variable
}
