% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{site.load}
\alias{site.load}
\title{Function to load/get a page on a site from the Qd server
The function requires the name of the page
The data is sent back in JSON.}
\usage{
site.load(page_name, ...)
}
\arguments{
\item{page_name}{name of the page}

\item{...}{}

\item{data}{\itemize{
\item JSON data that needs to be sent
}}
}
\value{
return JSON data of the specific page from the Qd server
}
\description{
Function to load/get a page on a site from the Qd server
The function requires the name of the page
The data is sent back in JSON.
}
\examples{
site.load("/example")
}
