% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{site.unload}
\alias{site.unload}
\title{Function to unload a page on site on the Qd server
The function requires the name of a page. This page will
then be removed from the server
Currently, Qd, server does not allow removing files - BROKEN}
\usage{
site.unload(page_name, ...)
}
\arguments{
\item{page_name}{\itemize{
\item name of the page to be removed from the Qd server
}}

\item{...}{}
}
\value{
http code if the page was removed
}
\description{
Function to unload a page on site on the Qd server
The function requires the name of a page. This page will
then be removed from the server
Currently, Qd, server does not allow removing files - BROKEN
}
