% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{site.upload}
\alias{site.upload}
\title{Function to upload files to the Qd server
This function will return the relative file path on the Qd server.
You can use this relative path for the purposes of your site/page/card}
\usage{
site.upload(files_list, ...)
}
\arguments{
\item{files_list}{a list of files names that neds to be uploaded to the server}

\item{...}{}
}
\value{
list of file names with their location on the Qd server
}
\description{
Function to upload files to the Qd server
This function will return the relative file path on the Qd server.
You can use this relative path for the purposes of your site/page/card
}
