% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tian_method.R
\name{tian_ve}
\alias{tian_ve}
\title{Estimate VE using method from Tian et al. 2005}
\usage{
tian_ve(dat, n_days, n_periods, n_days_period, alpha = 0.05)
}
\arguments{
\item{dat}{data set}

\item{n_days}{number of days in the study}

\item{n_periods}{number of periods in the study}

\item{n_days_period}{number of days per period}

\item{alpha}{hypothesis critical value}
}
\value{
number of times the null hypothesis is rejected
}
\description{
In this method, kernel-weighted partial likelihood approach is used to estimate the time-dependent coefficient in the generalized Cox model [REF] .  At each time point, the estimate is obtained by maximizing a smooth concave function of a p x 1 vector of parameters, where p is the dimension of the vector of covariates.
}
\keyword{wave}
