% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_breadcrumb}
\alias{ui_breadcrumb}
\title{Create a breadcrumb for a \code{h2o_q.types.BreadcrumbsCard()}.}
\usage{
ui_breadcrumb(name, label)
}
\arguments{
\item{name}{The name of this item. Prefix the name with a '#' to trigger hash-change navigation.}

\item{label}{The label to display.}
}
\value{
A Breadcrumb instance.
}
\description{
Create a breadcrumb for a \code{h2o_q.types.BreadcrumbsCard()}.
}
