% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_breadcrumbs_card}
\alias{ui_breadcrumbs_card}
\title{Create a card containing breadcrumbs.
Breadcrumbs should be used as a navigational aid in your app or site.
They indicate the current page’s location within a hierarchy and help
the user understand where they are in relation to the rest of that hierarchy.
They also afford one-click access to higher levels of that hierarchy.
Breadcrumbs are typically placed, in horizontal form, under the masthead
or navigation of an experience, above the primary content area.}
\usage{
ui_breadcrumbs_card(box, items, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{items}{A list of \code{h2o_q.types.Breadcrumb} instances to display. See \code{h2o_q.ui.breadcrumb()}}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A BreadcrumbsCard instance.
}
\description{
Create a card containing breadcrumbs.
Breadcrumbs should be used as a navigational aid in your app or site.
They indicate the current page’s location within a hierarchy and help
the user understand where they are in relation to the rest of that hierarchy.
They also afford one-click access to higher levels of that hierarchy.
Breadcrumbs are typically placed, in horizontal form, under the masthead
or navigation of an experience, above the primary content area.
}
