% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_button}
\alias{ui_button}
\title{Create a button.}
\usage{
ui_button(
  name,
  label = NULL,
  caption = NULL,
  value = NULL,
  primary = NULL,
  disabled = NULL,
  link = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component. If the name is prefixed with a '#', the button sets the location hash to the name when clicked.}

\item{label}{The text displayed on the button.}

\item{caption}{The caption displayed below the label. Setting a caption renders a compound button.}

\item{value}{A value for this button. If a value is set, it is used for the button's submitted instead of a boolean True.}

\item{primary}{True if the button should be rendered as the primary button in the set.}

\item{disabled}{True if the button should be disabled.}

\item{link}{True if the button should be rendered as link text and not a standard button.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Button instance.
}
\description{
Buttons are best used to enable a user to commit a change or complete steps in a task.
They are typically found inside forms, dialogs, panels or pages.
An example of their usage is confirming the deletion of a file in a confirmation dialog.
}
\details{
When considering their place in a layout, contemplate the order in which a user will flow through the UI.
As an example, in a form, the individual will need to read and interact with the form fields before submitting
the form. Therefore, as a general rule, the button should be placed at the bottom of the UI container
which holds the related UI elements.

Buttons may be placed within a "buttons" component which will lay out the buttons horizontally, or used
individually and they will be stacked vertically.

While buttons can technically be used to navigate a user to another part of the experience, this is not
recommended unless that navigation is part of an action or their flow.
}
