% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_buttons}
\alias{ui_buttons}
\title{Create a set of buttons to be layed out horizontally.}
\usage{
ui_buttons(items)
}
\arguments{
\item{items}{The button in this set.}
}
\value{
A Buttons instance.
}
\description{
Create a set of buttons to be layed out horizontally.
}
