% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_checkbox}
\alias{ui_checkbox}
\title{Create a checkbox.}
\usage{
ui_checkbox(
  name,
  label = NULL,
  value = NULL,
  indeterminate = NULL,
  disabled = NULL,
  trigger = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the checkbox.}

\item{value}{True if selected, False if unselected.}

\item{indeterminate}{True if the selection is indeterminate (neither selected nor unselected).}

\item{disabled}{True if the checkbox is disabled.}

\item{trigger}{True if the form should be submitted when the checkbox value changes.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Checkbox instance.
}
\description{
A checkbox allows users to switch between two mutually exclusive options (checked or unchecked, on or off) through
a single click or tap. It can also be used to indicate a subordinate setting or preference when paired with another
component.
}
\details{
A checkbox is used to select or deselect action items. It can be used for a single item or for a list of multiple
items that a user can choose from. The component has two selection states: unselected and selected.

For a binary choice, the main difference between a checkbox and a toggle switch is that the checkbox is for status
and the toggle switch is for action.

Use multiple checkboxes for multi-select scenarios in which a user chooses one or more items from a group of
choices that are not mutually exclusive.
}
