% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_checklist}
\alias{ui_checklist}
\title{Create a set of checkboxes.
Use this for multi-select scenarios in which a user chooses one or more items from a group of
choices that are not mutually exclusive.}
\usage{
ui_checklist(name, label = NULL, values = NULL, choices = NULL, tooltip = NULL)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed above the component.}

\item{values}{The names of the selected choices.}

\item{choices}{The choices to be presented.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Checklist instance.
}
\description{
Create a set of checkboxes.
Use this for multi-select scenarios in which a user chooses one or more items from a group of
choices that are not mutually exclusive.
}
