% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_choice}
\alias{ui_choice}
\title{Create a choice for a checklist, choice group or dropdown.}
\usage{
ui_choice(name, label = NULL, disabled = NULL)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the component.}

\item{disabled}{True if the checkbox is disabled.}
}
\value{
A Choice instance.
}
\description{
Create a choice for a checklist, choice group or dropdown.
}
