% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_choice_group}
\alias{ui_choice_group}
\title{Create a choice group.
The choice group component, also known as radio buttons, let users select one option from two or more choices.
Each option is represented by one choice group button; a user can select only one choice group in a button group.}
\usage{
ui_choice_group(
  name,
  label = NULL,
  value = NULL,
  choices = NULL,
  required = NULL,
  trigger = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the component.}

\item{value}{The name of the selected choice.}

\item{choices}{The choices to be presented.}

\item{required}{True if this field is required.}

\item{trigger}{True if the form should be submitted when the selection changes.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A ChoiceGroup instance.
}
\description{
Choice groups emphasize all options equally, and that may draw more attention to the options than necessary.
Consider using other components, unless the options deserve extra attention from the user.
For example, if the default option is recommended for most users in most situations, use a dropdown instead.
}
\details{
If there are only two mutually exclusive options, combine them into a single Checkbox or Toggle switch.
For example, use a checkbox for "I agree" instead of choice group buttons for "I agree" and "I don't agree."
}
