% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_color_picker}
\alias{ui_color_picker}
\title{Create a color picker.}
\usage{
ui_color_picker(
  name,
  label = NULL,
  value = NULL,
  choices = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the component.}

\item{value}{The selected color (CSS-compatible string)}

\item{choices}{A list of colors (CSS-compatible strings) to limit color choices to.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A ColorPicker instance.
}
\description{
A date picker allows a user to pick a color value.
If the 'choices' parameter is set, a swatch picker is displayed instead of the standard color picker.
}
