% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_combobox}
\alias{ui_combobox}
\title{Create a combobox.}
\usage{
ui_combobox(
  name,
  label = NULL,
  placeholder = NULL,
  value = NULL,
  choices = NULL,
  error = NULL,
  disabled = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the component.}

\item{placeholder}{A string that provides a brief hint to the user as to what kind of information is expected in the field.}

\item{value}{The name of the selected choice.}

\item{choices}{The choices to be presented.}

\item{error}{Text to be displayed as an error below the text box.}

\item{disabled}{True if this field is disabled.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Combobox instance.
}
\description{
A combobox is a list in which the selected item is always visible, and the others are visible on demand by
clicking a drop-down button or by typing in the input.
They are used to simplify the design and make a choice within the UI.
}
\details{
When closed, only the selected item is visible.
When users click the drop-down button, all the options become visible.
To change the value, users open the list and click another value or use the arrow keys (up and down)
to select a new value.
When collapsed the user can select a new value by typing.
}
