% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_command}
\alias{ui_command}
\title{Create a command.}
\usage{
ui_command(
  name,
  label = NULL,
  caption = NULL,
  icon = NULL,
  items = NULL,
  data = NULL
)
}
\arguments{
\item{name}{An identifying name for this component. If the name is prefixed with a '#', the command sets the location hash to the name when executed.}

\item{label}{The text displayed for this command.}

\item{caption}{The caption for this command (typically a tooltip).}

\item{icon}{The icon to be displayed for this command.}

\item{items}{Sub-commands, if any}

\item{data}{Data associated with this command, if any.}
}
\value{
A Command instance.
}
\description{
Commands are typically displayed as context menu items or toolbar button.
}
