% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_component}
\alias{ui_component}
\title{Create a component.}
\usage{
ui_component(
  text = NULL,
  text_xl = NULL,
  text_l = NULL,
  text_m = NULL,
  text_s = NULL,
  text_xs = NULL,
  label = NULL,
  separator = NULL,
  progress = NULL,
  message_bar = NULL,
  textbox = NULL,
  checkbox = NULL,
  toggle = NULL,
  choice_group = NULL,
  checklist = NULL,
  dropdown = NULL,
  combobox = NULL,
  slider = NULL,
  spinbox = NULL,
  date_picker = NULL,
  color_picker = NULL,
  button = NULL,
  buttons = NULL,
  file_upload = NULL,
  table = NULL,
  link = NULL,
  tabs = NULL,
  expander = NULL,
  frame = NULL,
  picker = NULL
)
}
\arguments{
\item{text}{Text block.}

\item{text_xl}{Extra-large sized text block.}

\item{text_l}{Large sized text block.}

\item{text_m}{Medium sized text block.}

\item{text_s}{Small sized text block.}

\item{text_xs}{Extra-small sized text block.}

\item{label}{Label.}

\item{separator}{Separator.}

\item{progress}{Progress bar.}

\item{message_bar}{Message bar.}

\item{textbox}{Textbox.}

\item{checkbox}{Checkbox.}

\item{toggle}{Toggle.}

\item{choice_group}{Choice group.}

\item{checklist}{Checklist.}

\item{dropdown}{Dropdown.}

\item{combobox}{Combobox.}

\item{slider}{Slider.}

\item{spinbox}{Spinbox.}

\item{date_picker}{Date picker.}

\item{color_picker}{Color picker.}

\item{button}{Button.}

\item{buttons}{Button set.}

\item{file_upload}{File upload.}

\item{table}{Table.}

\item{link}{Link.}

\item{tabs}{Tabs.}

\item{expander}{Expander.}

\item{frame}{Frame}

\item{picker}{Picker}
}
\value{
A Component instance.
}
\description{
Create a component.
}
