% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_date_picker}
\alias{ui_date_picker}
\title{Create a date picker.}
\usage{
ui_date_picker(
  name,
  label = NULL,
  placeholder = NULL,
  value = NULL,
  disabled = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the component.}

\item{placeholder}{A string that provides a brief hint to the user as to what kind of information is expected in the field.}

\item{value}{The date value in YYYY-MM-DD format.}

\item{disabled}{True if this field is disabled.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A DatePicker instance.
}
\description{
A date picker allows a user to pick a date value.
}
