% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_dropdown}
\alias{ui_dropdown}
\title{Create a dropdown.}
\usage{
ui_dropdown(
  name,
  label = NULL,
  placeholder = NULL,
  value = NULL,
  values = NULL,
  choices = NULL,
  required = NULL,
  disabled = NULL,
  trigger = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the component.}

\item{placeholder}{A string that provides a brief hint to the user as to what kind of information is expected in the field.}

\item{value}{The name of the selected choice.}

\item{values}{The names of the selected choices. If this parameter is set, multiple selections will be allowed.}

\item{choices}{The choices to be presented.}

\item{required}{True if this is a required field.}

\item{disabled}{True if this field is disabled.}

\item{trigger}{True if the form should be submitted when the dropdown value changes.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Dropdown instance.
}
\description{
A dropdown is a list in which the selected item is always visible, and the others are visible on demand by clicking
a drop-down button. They are used to simplify the design and make a choice within the UI. When closed, only the
selected item is visible. When users click the drop-down button, all the options become visible.
}
\details{
To change the value, users open the list and click another value or use the arrow keys (up and down) to
select a new value.

Note: Use either the 'value' parameter or the 'values' parameter. Setting the 'values' parameter renders a
multi-select dropdown.
}
