% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_expander}
\alias{ui_expander}
\title{Creates a new expander.}
\usage{
ui_expander(name, label = NULL, expanded = NULL, items = NULL)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{The text displayed on the expander.}

\item{expanded}{True if expanded, False if collapsed.}

\item{items}{List of components to be hideable by the expander.}
}
\value{
A Expander instance.
}
\description{
Expanders can be used to show or hide a group of related components.
}
