% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_file_upload}
\alias{ui_file_upload}
\title{Create a file upload component.
A file upload component allows a user to browse, select and upload one or more files.}
\usage{
ui_file_upload(
  name,
  label = NULL,
  multiple = NULL,
  file_extensions = NULL,
  max_file_size = NULL,
  max_size = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the component.}

\item{multiple}{True if the component should allow multiple files to be uploaded.}

\item{file_extensions}{List of allowed file extensions, e.g. \code{pdf}, \code{docx}, etc.}

\item{max_file_size}{Maximum allowed size (Mb) per file. Defaults to no limit.}

\item{max_size}{Maximum allowed size (Mb) for all files combined. Defaults to no limit.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A FileUpload instance.
}
\description{
Create a file upload component.
A file upload component allows a user to browse, select and upload one or more files.
}
