% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_flex_card}
\alias{ui_flex_card}
\title{EXPERIMENTAL. DO NOT USE.
Create a card containing other cards laid out using a one-dimensional model with flexible alignemnt and wrapping capabilities.}
\usage{
ui_flex_card(
  box,
  item_view,
  item_props,
  data,
  direction = NULL,
  justify = NULL,
  align = NULL,
  wrap = NULL,
  commands = NULL
)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{item_view}{The child card type.}

\item{item_props}{The child card properties.}

\item{data}{Data for this card.}

\item{direction}{Layout direction.
One of 'horizontal', 'vertical'.}

\item{justify}{Layout strategy for main axis.
One of 'start', 'end', 'center', 'between', 'around'.}

\item{align}{Layout strategy for cross axis.
One of 'start', 'end', 'center', 'baseline', 'stretch'.}

\item{wrap}{Wrapping strategy.
One of 'start', 'end', 'center', 'between', 'around', 'stretch'.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A FlexCard instance.
}
\description{
EXPERIMENTAL. DO NOT USE.
Create a card containing other cards laid out using a one-dimensional model with flexible alignemnt and wrapping capabilities.
}
