% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_form_card}
\alias{ui_form_card}
\title{Create a form.}
\usage{
ui_form_card(box, items, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{items}{The components in this form.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A FormCard instance.
}
\description{
Create a form.
}
