% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_frame}
\alias{ui_frame}
\title{Create a new inline frame (an \code{iframe}).}
\usage{
ui_frame(path = NULL, content = NULL, width = NULL, height = NULL)
}
\arguments{
\item{path}{The path or URL of the web page, e.g. \verb{/foo.html} or \verb{http://example.com/foo.html}}

\item{content}{The HTML content of the page. A string containing \verb{<html>...</html>}.}

\item{width}{The width of the frame, e.g. \verb{200px}, \verb{50\%}, etc. Defaults to \verb{100\%}.}

\item{height}{The height of the frame, e.g. \verb{200px}, \verb{50\%}, etc. Defaults to \verb{150px}.}
}
\value{
A Frame instance.
}
\description{
Create a new inline frame (an \code{iframe}).
}
