% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_frame_card}
\alias{ui_frame_card}
\title{Render a card containing a HTML page inside an inline frame (an \code{iframe}).}
\usage{
ui_frame_card(box, title, path = NULL, content = NULL, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The title for this card.}

\item{path}{The path or URL of the web page, e.g. \verb{/foo.html} or \verb{http://example.com/foo.html}}

\item{content}{The HTML content of the page. A string containing \verb{<html>...</html>}}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A FrameCard instance.
}
\description{
Either a path or content can be provided as arguments.
}
