% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_graphics_card}
\alias{ui_graphics_card}
\title{Create a card for displaying vector graphics.}
\usage{
ui_graphics_card(
  box,
  view_box,
  stage = NULL,
  scene = NULL,
  width = NULL,
  height = NULL,
  commands = NULL
)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{view_box}{The position and dimension of the SVG viewport, in user space.
A space-separated list of four numbers: min-x, min-y, width and height.
For example, '0 0 400 300'.
See: https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/viewBox}

\item{stage}{Background layer for rendering static SVG elements. Must be packed to conserve memory.}

\item{scene}{Foreground layer for rendering dynamic SVG elements.}

\item{width}{The displayed width of the rectangular viewport.
(Not the width of its coordinate system.)}

\item{height}{The displayed height of the rectangular viewport.
(Not the height of its coordinate system.)}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A GraphicsCard instance.
}
\description{
Create a card for displaying vector graphics.
}
