% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_grid_card}
\alias{ui_grid_card}
\title{EXPERIMENTAL. DO NOT USE.}
\usage{
ui_grid_card(box, title, cells, data, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{EXPERIMENTAL. DO NOT USE.}

\item{cells}{EXPERIMENTAL. DO NOT USE.}

\item{data}{EXPERIMENTAL. DO NOT USE.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A GridCard instance.
}
\description{
EXPERIMENTAL. DO NOT USE.
}
