% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_header_card}
\alias{ui_header_card}
\title{Render a card containing a HTML page inside an inline frame (iframe).}
\usage{
ui_header_card(
  box,
  title,
  subtitle,
  icon = NULL,
  icon_color = NULL,
  commands = NULL
)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The title.}

\item{subtitle}{The subtitle, displayed below the title.}

\item{icon}{The icon type, displayed to the left.}

\item{icon_color}{The icon's color.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A HeaderCard instance.
}
\description{
Either a path or content can be provided as arguments.
}
