% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_image_card}
\alias{ui_image_card}
\title{Create a card that displays a base64-encoded image.}
\usage{
ui_image_card(box, title, type, image, data = NULL, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The card's title.}

\item{type}{The image MIME subtype. One of \code{apng}, \code{bmp}, \code{gif}, \code{x-icon}, \code{jpeg}, \code{png}, \code{webp}.}

\item{image}{Image data, base64-encoded.}

\item{data}{Data for this card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A ImageCard instance.
}
\description{
Create a card that displays a base64-encoded image.
}
