% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_label}
\alias{ui_label}
\title{Create a label.}
\usage{
ui_label(label, required = NULL, disabled = NULL, tooltip = NULL)
}
\arguments{
\item{label}{The text displayed on the label.}

\item{required}{True if the field is required.}

\item{disabled}{True if the label should be disabled.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Label instance.
}
\description{
Labels give a name or title to a component or group of components.
Labels should be in close proximity to the component or group they are paired with.
Some components, such as textboxes, dropdowns, or toggles, already have labels
incorporated, but other components may optionally add a Label if it helps inform
the user of the component’s purpose.
}
