% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_large_bar_stat_card}
\alias{ui_large_bar_stat_card}
\title{Create a large captioned card displaying a primary value, an auxiliary value and a progress bar, with captions for each value.}
\usage{
ui_large_bar_stat_card(
  box,
  title,
  caption,
  value,
  aux_value,
  value_caption,
  aux_value_caption,
  progress,
  plot_color = NULL,
  data = NULL,
  commands = NULL
)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The card's title.}

\item{caption}{The card's caption.}

\item{value}{The primary value displayed.}

\item{aux_value}{The auxiliary value, typically a target value.}

\item{value_caption}{The caption displayed below the primary value.}

\item{aux_value_caption}{The caption displayed below the auxiliary value.}

\item{progress}{The value of the progress bar, between 0 and 1.}

\item{plot_color}{The color of the progress bar.}

\item{data}{Data for this card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A LargeBarStatCard instance.
}
\description{
Create a large captioned card displaying a primary value, an auxiliary value and a progress bar, with captions for each value.
}
