% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_large_stat_card}
\alias{ui_large_stat_card}
\title{Create a stat card displaying a primary value, an auxiliary value and a caption.}
\usage{
ui_large_stat_card(
  box,
  title,
  value,
  aux_value,
  caption,
  data = NULL,
  commands = NULL
)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The card's title.}

\item{value}{The primary value displayed.}

\item{aux_value}{The auxiliary value displayed next to the primary value.}

\item{caption}{The caption displayed below the primary value.}

\item{data}{Data for this card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A LargeStatCard instance.
}
\description{
Create a stat card displaying a primary value, an auxiliary value and a caption.
}
