% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_link}
\alias{ui_link}
\title{Create a hyperlink.}
\usage{
ui_link(
  label = NULL,
  path = NULL,
  disabled = NULL,
  download = NULL,
  button = NULL,
  tooltip = NULL
)
}
\arguments{
\item{label}{The text to be displayed. If blank, the \code{path} is used as the label.}

\item{path}{The path or URL to link to.}

\item{disabled}{True if the link should be disabled.}

\item{download}{True if the link should be used for file download.}

\item{button}{True if the link should be rendered as a button.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Link instance.
}
\description{
Hyperlinks can be internal or external.
Internal hyperlinks have paths that begin with a \code{/} and point to URLs within the Q UI.
All other kinds of paths are treated as external hyperlinks.
}
