% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_list_card}
\alias{ui_list_card}
\title{EXPERIMENTAL. DO NOT USE.
Create a card containing other cards laid out in the form of a list (vertically, top-to-bottom).}
\usage{
ui_list_card(box, title, item_view, item_props, data, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The title for this card.}

\item{item_view}{The child card type.}

\item{item_props}{The child card properties.}

\item{data}{Data for this card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A ListCard instance.
}
\description{
EXPERIMENTAL. DO NOT USE.
Create a card containing other cards laid out in the form of a list (vertically, top-to-bottom).
}
