% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_list_item1_card}
\alias{ui_list_item1_card}
\title{EXPERIMENTAL. DO NOT USE.}
\usage{
ui_list_item1_card(
  box,
  title,
  caption,
  value,
  aux_value,
  data,
  commands = NULL
)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{EXPERIMENTAL. DO NOT USE.}

\item{caption}{EXPERIMENTAL. DO NOT USE.}

\item{value}{EXPERIMENTAL. DO NOT USE.}

\item{aux_value}{EXPERIMENTAL. DO NOT USE.}

\item{data}{EXPERIMENTAL. DO NOT USE.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A ListItem1Card instance.
}
\description{
EXPERIMENTAL. DO NOT USE.
}
