% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_mark}
\alias{ui_mark}
\title{Create a specification for a layer of graphical marks such as bars, lines, points for a plot.
A plot can contain multiple such layers of marks.}
\usage{
ui_mark(
  coord = NULL,
  type = NULL,
  x = NULL,
  x0 = NULL,
  x1 = NULL,
  x2 = NULL,
  x_min = NULL,
  x_max = NULL,
  x_nice = NULL,
  x_scale = NULL,
  x_title = NULL,
  y = NULL,
  y0 = NULL,
  y1 = NULL,
  y2 = NULL,
  y_min = NULL,
  y_max = NULL,
  y_nice = NULL,
  y_scale = NULL,
  y_title = NULL,
  color = NULL,
  color_range = NULL,
  shape = NULL,
  shape_range = NULL,
  size = NULL,
  size_range = NULL,
  stack = NULL,
  dodge = NULL,
  curve = NULL,
  fill_color = NULL,
  fill_opacity = NULL,
  stroke_color = NULL,
  stroke_opacity = NULL,
  stroke_size = NULL,
  stroke_dash = NULL,
  label = NULL,
  label_offset = NULL,
  label_offset_x = NULL,
  label_offset_y = NULL,
  label_rotation = NULL,
  label_position = NULL,
  label_overlap = NULL,
  label_fill_color = NULL,
  label_fill_opacity = NULL,
  label_stroke_color = NULL,
  label_stroke_opacity = NULL,
  label_stroke_size = NULL,
  label_font_size = NULL,
  label_font_weight = NULL,
  label_line_height = NULL,
  label_align = NULL,
  ref_stroke_color = NULL,
  ref_stroke_opacity = NULL,
  ref_stroke_size = NULL,
  ref_stroke_dash = NULL
)
}
\arguments{
\item{coord}{Coordinate system. \code{rect} is synonymous to \code{cartesian}. \code{theta} is transposed \code{polar}.
One of 'rect', 'cartesian', 'polar', 'theta', 'helix'.}

\item{type}{Graphical geometry.
One of 'interval', 'line', 'path', 'point', 'area', 'polygon', 'schema', 'edge', 'heatmap'.}

\item{x}{X field or value.}

\item{x0}{X base field or value.}

\item{x1}{X bin lower bound field or value. For histograms.}

\item{x2}{X bin upper bound field or value. For histograms.}

\item{x_min}{X axis scale minimum.}

\item{x_max}{X axis scale maximum.}

\item{x_nice}{Whether to nice X axis scale ticks.}

\item{x_scale}{X axis scale type.
One of 'linear', 'cat', 'category', 'identity', 'log', 'pow', 'time', 'timeCat', 'quantize', 'quantile'.}

\item{x_title}{X axis title.}

\item{y}{Y field or value.}

\item{y0}{Y base field or value.}

\item{y1}{Y bin lower bound field or value. For histograms.}

\item{y2}{Y bin upper bound field or value. For histograms.}

\item{y_min}{Y axis scale minimum.}

\item{y_max}{Y axis scale maximum.}

\item{y_nice}{Whether to nice Y axis scale ticks.}

\item{y_scale}{Y axis scale type.
One of 'linear', 'cat', 'category', 'identity', 'log', 'pow', 'time', 'timeCat', 'quantize', 'quantile'.}

\item{y_title}{Y axis title.}

\item{color}{Mark color field or value.}

\item{color_range}{Mark color range for multi-series plots. A string containing space-separated colors, e.g. \code{'#fee8c8 #fdbb84 #e34a33'}}

\item{shape}{Mark shape field or value for \code{point} mark types. Possible values are 'circle', 'square', 'bowtie', 'diamond', 'hexagon', 'triangle', 'triangle-down', 'cross', 'tick', 'plus', 'hyphen', 'line'.}

\item{shape_range}{Mark shape range for multi-series plots using \code{point} mark types. A string containing space-separated shapes, e.g. \code{'circle square diamond'}}

\item{size}{Mark size field or value.}

\item{size_range}{Mark size range. A string containing space-separated integers, e.g. \code{'4 30'}}

\item{stack}{Field to stack marks by, or 'auto' to infer.}

\item{dodge}{Field to dodge marks by, or 'auto' to infer.}

\item{curve}{Curve type for \code{line} and \code{area} mark types.
One of 'none', 'smooth', 'step-before', 'step', 'step-after'.}

\item{fill_color}{Mark fill color.}

\item{fill_opacity}{Mark fill opacity.}

\item{stroke_color}{Mark stroke color.}

\item{stroke_opacity}{Mark stroke opacity.}

\item{stroke_size}{Mark stroke size.}

\item{stroke_dash}{Mark stroke dash style. A string containing space-separated integers that specify distances to alternately draw a line and a gap (in coordinate space units). If the number of elements in the array is odd, the elements of the array get copied and concatenated. For example, \link{5, 15, 25} will become \link{5, 15, 25, 5, 15, 25}.}

\item{label}{Label field or value.}

\item{label_offset}{Distance between label and mark.}

\item{label_offset_x}{Horizontal distance between label and mark.}

\item{label_offset_y}{Vertical distance between label and mark.}

\item{label_rotation}{Label rotation angle, in degrees, or 'none' to disable automatic rotation. The default behavior is 'auto' for automatic rotation.}

\item{label_position}{Label position relative to the mark.
One of 'top', 'bottom', 'middle', 'left', 'right'.}

\item{label_overlap}{Strategy to use if labels overlap.
One of 'hide', 'overlap', 'constrain'.}

\item{label_fill_color}{Label fill color.}

\item{label_fill_opacity}{Label fill opacity.}

\item{label_stroke_color}{Label stroke color.}

\item{label_stroke_opacity}{Label stroke opacity.}

\item{label_stroke_size}{Label stroke size (line width or pen thickness).}

\item{label_font_size}{Label font size.}

\item{label_font_weight}{Label font weight.}

\item{label_line_height}{Label line height.}

\item{label_align}{Label text alignment.
One of 'left', 'right', 'center', 'start', 'end'.}

\item{ref_stroke_color}{Reference line stroke color.}

\item{ref_stroke_opacity}{Reference line stroke opacity.}

\item{ref_stroke_size}{Reference line stroke size (line width or pen thickness).}

\item{ref_stroke_dash}{Reference line stroke dash style. A string containing space-separated integers that specify distances to alternately draw a line and a gap (in coordinate space units). If the number of elements in the array is odd, the elements of the array get copied and concatenated. For example, \link{5, 15, 25} will become \link{5, 15, 25, 5, 15, 25}.}
}
\value{
A Mark instance.
}
\description{
Create a specification for a layer of graphical marks such as bars, lines, points for a plot.
A plot can contain multiple such layers of marks.
}
