% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_markdown_card}
\alias{ui_markdown_card}
\title{Create a card that renders Markdown content.}
\usage{
ui_markdown_card(box, title, content, data = NULL, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The title for this card.}

\item{content}{The markdown content. Supports Github Flavored Markdown (GFM): https://guides.github.com/features/mastering-markdown/}

\item{data}{Additional data for the card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A MarkdownCard instance.
}
\description{
Github-flavored markdown is supported.
HTML markup is allowed in markdown content.
URLs, if found, are displayed as hyperlinks.
Copyright, reserved, trademark, quotes, etc. are replaced with language-neutral symbols.
}
