% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_markup_card}
\alias{ui_markup_card}
\title{Render HTML content.}
\usage{
ui_markup_card(box, title, content, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The title for this card.}

\item{content}{The HTML content.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A MarkupCard instance.
}
\description{
Render HTML content.
}
