% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_message_bar}
\alias{ui_message_bar}
\title{Create a message bar.}
\usage{
ui_message_bar(type = NULL, text = NULL)
}
\arguments{
\item{type}{The icon and color of the message bar.
One of 'info', 'error', 'warning', 'success', 'danger', 'blocked'.}

\item{text}{The text displayed on the message bar.}
}
\value{
A MessageBar instance.
}
\description{
A message bar is an area at the top of a primary view that displays relevant status information.
You can use a message bar to tell the user about a situation that does not require their immediate attention and
therefore does not need to block other activities.
}
