% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_meta_card}
\alias{ui_meta_card}
\title{Represents page-global state.}
\usage{
ui_meta_card(
  box,
  title = NULL,
  refresh = NULL,
  notification = NULL,
  commands = NULL
)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The title of the page.}

\item{refresh}{Refresh rate in seconds. A value of 0 turns off live-updates. Values != 0 are currently ignored (reserved for future use).}

\item{notification}{Display a desktop notification to the user.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A MetaCard instance.
}
\description{
This card is invisible.
It is used to control attributes of the active page.
}
