% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_nav_card}
\alias{ui_nav_card}
\title{Create a card containing a navigation pane.}
\usage{
ui_nav_card(box, items, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{items}{The navigation groups contained in this pane.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A NavCard instance.
}
\description{
Create a card containing a navigation pane.
}
