% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_nav_group}
\alias{ui_nav_group}
\title{Create a group of navigation items.}
\usage{
ui_nav_group(label, items)
}
\arguments{
\item{label}{The label to display for this group.}

\item{items}{The navigation items contained in this group.}
}
\value{
A NavGroup instance.
}
\description{
Create a group of navigation items.
}
