% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_nav_item}
\alias{ui_nav_item}
\title{Create a navigation item.}
\usage{
ui_nav_item(name, label)
}
\arguments{
\item{name}{The name of this item. Prefix the name with a '#' to trigger hash-change navigation.}

\item{label}{The label to display.}
}
\value{
A NavItem instance.
}
\description{
Create a navigation item.
}
