% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_picker}
\alias{ui_picker}
\title{Create a picker.
Pickers are used to select one or more choices, such as tags or files, from a list.
Use a picker to allow the user to quickly search for or manage a few tags or files.}
\usage{
ui_picker(
  name,
  choices,
  label = NULL,
  values = NULL,
  max_choices = NULL,
  disabled = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{choices}{The choices to be presented.}

\item{label}{Text to be displayed above the component.}

\item{values}{The names of the selected choices.}

\item{max_choices}{Maximum number of selectable choices. Defaults to no limit.}

\item{disabled}{Controls whether the picker should be disabled or not.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Picker instance.
}
\description{
Create a picker.
Pickers are used to select one or more choices, such as tags or files, from a list.
Use a picker to allow the user to quickly search for or manage a few tags or files.
}
