% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_pixel_art_card}
\alias{ui_pixel_art_card}
\title{Create a card displaying a collaborative Pixel art tool, just for kicks.}
\usage{
ui_pixel_art_card(box, title, data, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The title for this card.}

\item{data}{The data for this card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A PixelArtCard instance.
}
\description{
Create a card displaying a collaborative Pixel art tool, just for kicks.
}
