% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_plot}
\alias{ui_plot}
\title{Create a plot. A plot is composed of one or more graphical mark layers.}
\usage{
ui_plot(marks)
}
\arguments{
\item{marks}{The graphical mark layers contained in this plot.}
}
\value{
A Plot instance.
}
\description{
Create a plot. A plot is composed of one or more graphical mark layers.
}
