% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_plot_card}
\alias{ui_plot_card}
\title{Create a card displaying a plot.}
\usage{
ui_plot_card(box, title, data, plot, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The title for this card.}

\item{data}{Data for this card.}

\item{plot}{The plot to be displayed in this card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A PlotCard instance.
}
\description{
Create a card displaying a plot.
}
