% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_progress}
\alias{ui_progress}
\title{Create a progress bar.}
\usage{
ui_progress(label, caption = NULL, value = NULL, tooltip = NULL)
}
\arguments{
\item{label}{The text displayed above the bar.}

\item{caption}{The text displayed below the bar.}

\item{value}{The progress, between 0.0 and 1.0, or -1 (default) if indeterminate.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Progress instance.
}
\description{
Progress bars are used to show the completion status of an operation lasting more than 2 seconds.
If the state of progress cannot be determined, do not set a value.
Progress bars feature a bar showing total units to completion, and total units finished.
The label appears above the bar, and the caption appears below.
The label should tell someone exactly what the operation is doing.
}
\details{
Examples of formatting include:
\link{Object} is being \link{operation name}, or
\link{Object} is being \link{operation name} to \link{destination name} or
\link{Object} is being \link{operation name} from \link{source name} to \link{destination name}

Status text is generally in units elapsed and total units.
Real-world examples include copying files to a storage location, saving edits to a file, and more.
Use units that are informative and relevant to give the best idea to users of how long the operation will take to complete.
Avoid time units as they are rarely accurate enough to be trustworthy.
Also, combine steps of a complex operation into one total bar to avoid “rewinding” the bar.
Instead change the label to reflect the change if necessary. Bars moving backwards reduce confidence in the service.
}
