% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_repeat_card}
\alias{ui_repeat_card}
\title{EXPERIMENTAL. DO NOT USE.
Create a card containing other cards.}
\usage{
ui_repeat_card(box, item_view, item_props, data, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{item_view}{EXPERIMENTAL. DO NOT USE.}

\item{item_props}{The child card properties.}

\item{data}{Data for this card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A RepeatCard instance.
}
\description{
EXPERIMENTAL. DO NOT USE.
Create a card containing other cards.
}
