% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_separator}
\alias{ui_separator}
\title{Create a separator.}
\usage{
ui_separator(label = NULL)
}
\arguments{
\item{label}{The text displayed on the separator.}
}
\value{
A Separator instance.
}
\description{
A separator visually separates content into groups.
}
