% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_slider}
\alias{ui_slider}
\title{Create a slider.}
\usage{
ui_slider(
  name,
  label = NULL,
  min = NULL,
  max = NULL,
  step = NULL,
  value = NULL,
  disabled = NULL,
  trigger = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the component.}

\item{min}{The minimum value of the slider.}

\item{max}{The maximum value of the slider.}

\item{step}{The difference between two adjacent values of the slider.}

\item{value}{The current value of the slider.}

\item{disabled}{True if this field is disabled.}

\item{trigger}{True if the form should be submitted when the slider value changes.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Slider instance.
}
\description{
A slider is an element used to set a value. It provides a visual indication of adjustable content, as well as the
current setting in the total range of content. It is displayed as a horizontal track with options on either side.
A knob or lever is dragged to one end or the other to make the choice, indicating the current value.
Marks on the slider bar can show values and users can choose where they want to drag the knob or lever to
set the value.
}
\details{
A slider is a good choice when you know that users think of the value as a relative quantity, not a numeric value.
For example, users think about setting their audio volume to low or medium — not about setting the
value to two or five.

The default value of the slider will be zero or be constrained to the min and max values. The min will be returned
if the value is set under the min and the max will be returned if set higher than the max value.
}
