% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_small_stat_card}
\alias{ui_small_stat_card}
\title{Create a stat card displaying a single value.}
\usage{
ui_small_stat_card(box, title, value, data = NULL, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The card's title.}

\item{value}{The primary value displayed.}

\item{data}{Data for this card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A SmallStatCard instance.
}
\description{
Create a stat card displaying a single value.
}
