% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_spinbox}
\alias{ui_spinbox}
\title{Create a spinbox.}
\usage{
ui_spinbox(
  name,
  label = NULL,
  min = NULL,
  max = NULL,
  step = NULL,
  value = NULL,
  disabled = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the component.}

\item{min}{The minimum value of the spinbox.}

\item{max}{The maximum value of the spinbox.}

\item{step}{The difference between two adjacent values of the spinbox.}

\item{value}{The current value of the spinbox.}

\item{disabled}{True if this field is disabled.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Spinbox instance.
}
\description{
A spinbox allows the user to incrementally adjust a value in small steps.
It is mainly used for numeric values, but other values are supported too.
}
