% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_tab}
\alias{ui_tab}
\title{Create a tab.}
\usage{
ui_tab(name, label = NULL, icon = NULL)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{The text displayed on the tab.}

\item{icon}{The icon displayed on the tab.}
}
\value{
A Tab instance.
}
\description{
Create a tab.
}
