% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_tab_card}
\alias{ui_tab_card}
\title{Create a card containing tabs for navigation.}
\usage{
ui_tab_card(box, items, link = NULL, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{items}{Items to render.}

\item{link}{True if tabs should be rendered as links and not a standard tab.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A TabCard instance.
}
\description{
Create a card containing tabs for navigation.
}
