% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_table}
\alias{ui_table}
\title{Create an interactive table.}
\usage{
ui_table(name, columns, rows, multiple = NULL, tooltip = NULL)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{columns}{The columns in this table.}

\item{rows}{The rows in this table.}

\item{multiple}{True to allow multiple rows to be selected.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Table instance.
}
\description{
This table differs from a markdown table in that it supports clicking or selecting rows. If you simply want to
display a non-interactive table of information, use a markdown table.
}
\details{
If \code{multiple} is set to False (default), each row in the table is clickable. When a row is clicked, the form is
submitted automatically, and \code{q.args.table_name} is set to \verb{[row_name]}, where \code{table_name} is the \code{name} of
the table, and \code{row_name} is the \code{name} of the row that was clicked on.

If \code{multiple} is set to True, each row in the table is selectable. A row can be selected by clicking on it.
Multiple rows can be selected either by shift+clicking or using marquee selection. When the form is submitted,
\code{q.args.table_name} is set to \verb{[row1_name, row2_name, ...]} where \code{table_name} is the \code{name} of the table,
and \code{row1_name}, \code{row2_name} are the \code{name} of the rows that were selected. Note that if \code{multiple} is
set to True, the form is not submitted automatically, and one or more buttons are required in the form to trigger
submission.
}
