% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_table_column}
\alias{ui_table_column}
\title{Create a table column.}
\usage{
ui_table_column(name, label)
}
\arguments{
\item{name}{An identifying name for this column.}

\item{label}{The text displayed on the column header.}
}
\value{
A TableColumn instance.
}
\description{
Create a table column.
}
