% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_table_row}
\alias{ui_table_row}
\title{Create a table row.}
\usage{
ui_table_row(name, cells)
}
\arguments{
\item{name}{An identifying name for this row.}

\item{cells}{The cells in this row (displayed left to right).}
}
\value{
A TableRow instance.
}
\description{
Create a table row.
}
