% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_tabs}
\alias{ui_tabs}
\title{Create a tab bar.}
\usage{
ui_tabs(name, value = NULL, items = NULL)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{value}{The name of the tab to select.}

\item{items}{The tabs in this tab bar.}
}
\value{
A Tabs instance.
}
\description{
Create a tab bar.
}
