% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_tall_series_stat_card}
\alias{ui_tall_series_stat_card}
\title{Create a tall stat card displaying a primary value, an auxiliary value and a series plot.}
\usage{
ui_tall_series_stat_card(
  box,
  title,
  value,
  aux_value,
  plot_data,
  plot_value,
  plot_zero_value = NULL,
  plot_category = NULL,
  plot_type = NULL,
  plot_curve = NULL,
  plot_color = NULL,
  data = NULL,
  commands = NULL
)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The card's title.}

\item{value}{The primary value displayed.}

\item{aux_value}{The auxiliary value displayed below the primary value.}

\item{plot_data}{The plot's data.}

\item{plot_value}{The data field to use for y-axis values.}

\item{plot_zero_value}{The base value to use for each y-axis mark. Set this to \code{0} if you want to pin the x-axis at \code{y=0}. If not provided, the minimum value from the data is used.}

\item{plot_category}{The data field to use for x-axis values (ignored if \code{plot_type} is \code{area}; must be provided if \code{plot_type} is \code{interval}). Defaults to 'x'.}

\item{plot_type}{The type of plot. Defaults to \code{area}.
One of 'area', 'interval'.}

\item{plot_curve}{The plot's curve style. Defaults to \code{linear}.
One of 'linear', 'smooth', 'step', 'step-after', 'step-before'.}

\item{plot_color}{The plot's color.}

\item{data}{Data for this card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A TallSeriesStatCard instance.
}
\description{
Create a tall stat card displaying a primary value, an auxiliary value and a series plot.
}
