% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_template_card}
\alias{ui_template_card}
\title{Render dynamic content using a HTML template.}
\usage{
ui_template_card(box, title, content, data = NULL, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The title for this card.}

\item{content}{The Handlebars template. https://handlebarsjs.com/guide/}

\item{data}{Data for the Handlebars template}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A TemplateCard instance.
}
\description{
Render dynamic content using a HTML template.
}
