% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_text}
\alias{ui_text}
\title{Create text content.}
\usage{
ui_text(content, size = NULL, tooltip = NULL)
}
\arguments{
\item{content}{The text content.}

\item{size}{The font size of the text content.
One of 'xl', 'l', 'm', 's', 'xs'.}

\item{tooltip}{Tooltip message.}
}
\value{
A Text instance.
}
\description{
Create text content.
}
