% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_text_l}
\alias{ui_text_l}
\title{Create large sized text content.}
\usage{
ui_text_l(content, tooltip = NULL)
}
\arguments{
\item{content}{The text content.}

\item{tooltip}{Tooltip message.}
}
\value{
A TextL instance.
}
\description{
Create large sized text content.
}
