% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_text_m}
\alias{ui_text_m}
\title{Create medium sized text content.}
\usage{
ui_text_m(content, tooltip = NULL)
}
\arguments{
\item{content}{The text content.}

\item{tooltip}{Tooltip message.}
}
\value{
A TextM instance.
}
\description{
Create medium sized text content.
}
