% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_text_s}
\alias{ui_text_s}
\title{Create small sized text content.}
\usage{
ui_text_s(content, tooltip = NULL)
}
\arguments{
\item{content}{The text content.}

\item{tooltip}{Tooltip message.}
}
\value{
A TextS instance.
}
\description{
Create small sized text content.
}
