% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_text_xl}
\alias{ui_text_xl}
\title{Create extra-large sized text content.}
\usage{
ui_text_xl(content, tooltip = NULL)
}
\arguments{
\item{content}{The text content.}

\item{tooltip}{Tooltip message.}
}
\value{
A TextXl instance.
}
\description{
Create extra-large sized text content.
}
