% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_text_xs}
\alias{ui_text_xs}
\title{Create extra-small sized text content.}
\usage{
ui_text_xs(content, tooltip = NULL)
}
\arguments{
\item{content}{The text content.}

\item{tooltip}{Tooltip message.}
}
\value{
A TextXs instance.
}
\description{
Create extra-small sized text content.
}
