% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_textbox}
\alias{ui_textbox}
\title{Create a text box.}
\usage{
ui_textbox(
  name,
  label = NULL,
  placeholder = NULL,
  value = NULL,
  mask = NULL,
  icon = NULL,
  prefix = NULL,
  suffix = NULL,
  error = NULL,
  required = NULL,
  disabled = NULL,
  readonly = NULL,
  multiline = NULL,
  password = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{The text displayed above the field.}

\item{placeholder}{A string that provides a brief hint to the user as to what kind of information is expected in the field. It should be a word or short phrase that demonstrates the expected type of data, rather than an explanatory message.}

\item{value}{Text to be displayed inside the text box.}

\item{mask}{The masking string that defines the mask's behavior. A backslash will escape any character. Special format characters are: '9': \link{0-9} 'a': \link{a-zA-Z} '*': \link{a-zA-Z0-9}.}

\item{icon}{Icon displayed in the far right end of the text field.}

\item{prefix}{Text to be displayed before the text box contents.}

\item{suffix}{Text to be displayed after the text box contents.}

\item{error}{Text to be displayed as an error below the text box.}

\item{required}{True if the text box is a required field.}

\item{disabled}{True if the text box is disabled.}

\item{readonly}{True if the text box is a read-only field.}

\item{multiline}{True if the text box should allow multi-line text entry.}

\item{password}{True if the text box should hide text content.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Textbox instance.
}
\description{
The text box component enables a user to type text into an app.
It's typically used to capture a single line of text, but can be configured to capture multiple lines of text.
The text displays on the screen in a simple, uniform format.
}
