% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_toggle}
\alias{ui_toggle}
\title{Create a toggle.
Toggles represent a physical switch that allows users to turn things on or off.
Use toggles to present users with two mutually exclusive options (like on/off), where choosing an option results
in an immediate action.}
\usage{
ui_toggle(
  name,
  label = NULL,
  value = NULL,
  disabled = NULL,
  trigger = NULL,
  tooltip = NULL
)
}
\arguments{
\item{name}{An identifying name for this component.}

\item{label}{Text to be displayed alongside the component.}

\item{value}{True if selected, False if unselected.}

\item{disabled}{True if the checkbox is disabled.}

\item{trigger}{True if the form should be submitted when the toggle value changes.}

\item{tooltip}{An optional tooltip message displayed when a user clicks the help icon to the right of the component.}
}
\value{
A Toggle instance.
}
\description{
Use a toggle for binary operations that take effect right after the user flips the Toggle.
For example, use a Toggle to turn services or hardware components on or off.
In other words, if a physical switch would work for the action, a Toggle is probably the best component to use.
}
