% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_toolbar_card}
\alias{ui_toolbar_card}
\title{Create a card containing a toolbar.}
\usage{
ui_toolbar_card(
  box,
  items,
  secondary_items = NULL,
  overflow_items = NULL,
  commands = NULL
)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{items}{Items to render.}

\item{secondary_items}{Items to render on the right side (or left, in RTL).}

\item{overflow_items}{Items to render in an overflow menu.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A ToolbarCard instance.
}
\description{
Create a card containing a toolbar.
}
