% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_vega_card}
\alias{ui_vega_card}
\title{Create a card containing a Vega-lite plot.}
\usage{
ui_vega_card(box, title, specification, data = NULL, commands = NULL)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The title of this card.}

\item{specification}{The Vega-lite specification.}

\item{data}{Data for the plot, if any.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A VegaCard instance.
}
\description{
Create a card containing a Vega-lite plot.
}
