% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_wide_gauge_stat_card}
\alias{ui_wide_gauge_stat_card}
\title{Create a wide stat card displaying a primary value, an auxiliary value and a progress gauge.}
\usage{
ui_wide_gauge_stat_card(
  box,
  title,
  value,
  aux_value,
  progress,
  plot_color = NULL,
  data = NULL,
  commands = NULL
)
}
\arguments{
\item{box}{A string indicating how to place this component on the page.}

\item{title}{The card's title.}

\item{value}{The primary value displayed.}

\item{aux_value}{The auxiliary value displayed next to the primary value.}

\item{progress}{The value of the progress gauge, between 0 and 1.}

\item{plot_color}{The color of the progress gauge.}

\item{data}{Data for this card.}

\item{commands}{Contextual menu commands for this component.}
}
\value{
A WideGaugeStatCard instance.
}
\description{
Create a wide stat card displaying a primary value, an auxiliary value and a progress gauge.
}
