% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_methods.R
\name{ainslie_ve}
\alias{ainslie_ve}
\title{Estimate VE using a maximum likelihood (ML) method}
\usage{
ainslie_ve(
  dat,
  n_days,
  n_periods,
  n_days_period,
  latent_period = 1,
  infectious_period = 4
)
}
\arguments{
\item{dat}{data set}

\item{n_days}{number of days in the study}

\item{n_periods}{number of periods in the study}

\item{n_days_period}{number of days per period}

\item{latent_period}{length of latent period}

\item{infectious_period}{length of infectious period}
}
\value{
list with a tibble of VE estimates for each period (the estimate for each period is the average VE over the days
within each period) and the maximu likelihood estimates of the parameters.
}
\description{
This method is similar to that of Ainslie et al. 2017 (SIM). The ML method is based on calculating the
contribution to the likelihood function of each study participant. Let t (t=1,2,…T) denote the number
of days since vaccination or receipt of the placebo. Denote lambda_d = hazard of infection of an unvaccinated
susceptible person on day t. P(t) denotes the probability that a randomly selected population member
is infectious on day t  (prevalence of infection). We assume that the hazard of infection is proportional
to the prevalence: lambda_t=lambda∙P(t). A method for estimatingon the prevalence is described later. Next, let theta_t
be defined such that the hazard of a vaccinated susceptible person on day t is lambda_t∙theta_t. Then the TVE
on day t, defined as one minus the ratio of the hazards of infection for an unvaccinated and a
vaccinated person on that day is [1-theta]_t. Next, we define a status variable for each study participant
i (i=1,2,…N) on each day t:
* Y_it=0 if this person remained susceptible by the end of the day,
* Y_it=1 if this person became infected on this day, and
* Y_it=2 if this person became infected before this day.
(A person with Y_it=2 may still be infectious to others; we use this notation to indicate that this person
is no longer at risk of becoming infected). We assume that on day 0 everyone is susceptible, i.e., Y_i0=0.
}
\details{
Next, let  pi_itj=Pra({Y_it=j|Y_i(t-1) =0)}  be the conditional probability that person i is in status j
(j=0,1,2) given that s/he was susceptible at the end of the previous day. If person i is unvaccinated then
pi_it0=1-lambda_t,  pi_it1=lambda_t,  pi_it2=0. If person i is vaccinated then lambda_t in the last equations is replaced by
lambda_t∙theta_t . Denoting by psi_itj=Pra(a{Y_it=j)} the unconditional probabilities of Y_it, it is easy to see tha
t psi_it0= psi_(i(t-1)0)∙ pi_it0, psi_it1=psi_(i(t-1)0)∙ pi_it1,  psi_it2= 1-(psi_it0+psi_it1).

We now can write the contribution of each study subject to the likelihood function in terms of the
unconditional probabilities psi_itj. The contribution of a person who became infected on day t  is
[psi]_it1,  while the contribution of a person who did not become infected during the entire study
is psi_iT0, where T is the last day of the study. For a participant who dropped out uninfected, T is replaced
by the day s/he dropped out. We can assume that the contributions of different study participants are
independent because we condition on the daily prevalences when we determine the unconditional distribution
of each Y_it. Therefore, the final likelihood function is the product of the contributions of all the study
participants.

To evaluate waning of TVE, we recall that TVE on day t is [1-theta]_t.  Therefore, temporal changes in TVE
are determined by how theta_t depends on t. A simple approach to modeling TVE as a function of time assumes a
fixed daily increase phi≥0 in theta_t (which amounts to a fixed daily decrease in TVE), i.e. theta_t=theta_1+phi∙(t-1)
for t=1,2,…T. Then the likelihood function has 3 parameters: alpha,theta_1,and phi. One can calculate the maximum
likelihood estimated of these parameters along with their standard errors by maximizing the likelihood
function. One can then test the hypothesis phi=0 to determine the presence of waning and obtain a confidence
interval for phi, the daily rate of TVE waning.
}
\keyword{wave}
