% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_methods.R
\name{tian_ve}
\alias{tian_ve}
\title{Estimate VE using method from Tian et al. 2005}
\usage{
tian_ve(dat, n_days, n_periods, n_days_period, alpha = 0.05)
}
\arguments{
\item{dat}{data set}

\item{n_days}{number of days in the study}

\item{n_periods}{number of periods in the study}

\item{n_days_period}{number of days per period}

\item{alpha}{hypothesis critical value}
}
\value{
number of times the null hypothesis is rejected
}
\description{
In this method, kernel-weighted partial likelihood approach is used to estimate the time-dependent coefficient
in the generalized Cox model [REF] .  At each time point, the estimate is obtained by maximizing a smooth
concave function of a p x 1 vector of parameters, where p is the dimension of the vector of covariates. The
(1-alpha) confidence bands for the time-dependent coefficient beta(t) can be obtained by beta(t) ±c_alpha w [(t)]^(-1
), where t is in the time interval of interest (b_1≤t≤b_2), c_alpha is the 100(1-alpha)th percentile the approximate
distribution of beta(t)_hat, and w(t)_hat is a positive weighting function. To test the hypothesis of no VE waning
(i.e., the proportional hazards assumption is met), confidence bands from the distribution of the estimated
cumulative function of beta(t)_hat) can be obtained. If the line (0,0) is not contained within the confidence band
s over the entire time interval, then the proportional hazards assumption is not met and the null hypothesis
of no VE waning is rejected.
}
\keyword{wave}
